+++ Cargo.toml
@@ -51,7 +51,7 @@ members = [
     "tools/spicepodschema/",
     "tools/spiceschema",
     "tools/testoperator",
-    "tools/parsley",
+    "tools/parsley", "crates/runtime-datafusion",
 ]
 
 # Explicitly set the resolver to version 2, which is the default for packages with edition >= 2021
