+++ crates/runtime/src/accelerated_table/refresh_task.rs
@@ -30,6 +30,8 @@ use datafusion_table_providers::util::retriable_error::{
 };
 use futures::{StreamExt, stream};
 use opentelemetry::KeyValue;
+use runtime_datafusion::execution_plan::schema_cast::EnsureSchema;
+use runtime_datafusion::extension::bytes_processed::BytesProcessedExtensionPlanner;
 use runtime_datafusion_index::analyzer::{
     IndexTableScanExtensionPlanner, IndexTableScanOptimizerRule,
 };
@@ -41,19 +43,19 @@ use util::{RetryError, retry};
 
 use crate::datafusion::builder::{AnalyzerRulesBuilder, get_df_default_config};
 use crate::datafusion::error::{SpiceExternalError, find_datafusion_root, get_spice_df_error};
-use crate::datafusion::extension::{SpiceExtensionPlanner, SpiceQueryPlanner};
 use crate::datafusion::is_spice_internal_dataset;
 use crate::datafusion::schema::BaseSchema;
 use crate::federated_table::FederatedTable;
+use crate::metrics::telemetry::track_bytes_processed;
 use crate::timing::MultiTimeMeasurement;
 use crate::{
     component::dataset::acceleration::RefreshMode,
     dataconnector::get_data,
     datafusion::{filter_converter::TimestampFilterConvert, schema},
     dataupdate::{DataUpdate, StreamingDataUpdate, UpdateType},
-    execution_plan::schema_cast::EnsureSchema,
     status,
 };
+use runtime_datafusion::extension::ExtensionPlanQueryPlanner;
 use runtime_object_store::registry::default_runtime_env;
 
 use super::refresh::get_timestamp;
@@ -761,7 +763,9 @@ impl RefreshTask {
             .with_default_features();
 
         let mut extension_planners: Vec<Arc<dyn ExtensionPlanner + Send + Sync>> = vec![
-            Arc::new(SpiceExtensionPlanner::new()),
+            Arc::new(BytesProcessedExtensionPlanner::new(Box::new(
+                track_bytes_processed,
+            ))),
             Arc::new(IndexTableScanExtensionPlanner::new()),
         ];
 
@@ -775,7 +779,8 @@ impl RefreshTask {
             extension_planners.push(Arc::new(FederatedPlanner::new()));
         }
 
-        let query_planner = SpiceQueryPlanner::new().with_extension_planners(extension_planners);
+        let query_planner =
+            ExtensionPlanQueryPlanner::new().with_extension_planners(extension_planners);
 
         let mut state = state_builder
             .with_query_planner(Arc::new(query_planner))
